refactor: resolve #282 - split ProgramDB.test.ts from 663 to under 500 lines
Closes #282
